model/requirements.txt
fastapi==0.111.0
numpy==2.0.0
Pillow==10.3.0
skimage==0.0
tensorflow==2.16.1
uvicorn==0.30.1
